Moving stuff around
StorageBackend -> zinc/storages/__init__.py
ZincCatalog -> zinc/services/__init__.py
CatalogCoordinator -> zinc/coordinators/__init__.py
