+++ src/zinc/__init__.py
@@ -2,8 +2,6 @@
 from .errors import ZincErrors
 from .helpers import *
 from .defaults import defaults
-from .models import ZincIndex, ZincManifest, ZincFlavorSpec
-from .catalog import ZincCatalog, create_catalog_at_path
 
 import logging
 logging.basicConfig(level=logging.DEBUG,
